+++ GetThePicture/Picture/Clause/Base/Overpunch/OpCodec.cs
@@ -1,4 +1,5 @@
 using GetThePicture.Picture.Clause.Base.Options;
+using GetThePicture.Picture.Clause.Base.Overpunch.Codex;
 
 namespace GetThePicture.Picture.Clause.Base.Overpunch;
 
@@ -13,12 +14,12 @@ public static class OpCodec
     /// <param name="bytes"></param>
     /// <param name="pic"></param>
     /// <param name="options"></param>
-    /// <param name="sign">符號</param>
+    /// <param name="isNegative"></param>
     /// <returns></returns>
     /// <exception cref="FormatException"></exception>
-    public static Span<byte> Decode(Span<byte> bytes, PicMeta pic, CodecOptions options, out decimal sign)
+    public static Span<byte> Decode(Span<byte> bytes, PicMeta pic, CodecOptions options, out bool isNegative)
     {
-        sign = 1.0m;
+        isNegative = false;
 
         if (pic.Signed)
         {
@@ -30,11 +31,17 @@ public static class OpCodec
             };
 
             byte key = (byte)(bytes[index] & 0x7F); // ASCII overpunch
+            string opVal = OpCodex.TryGetValue(key, options.DataStorage);
 
-            OpVal opVal = GetOpValue(key, options.DataStorage);
+            char sign = opVal[0];
+            isNegative = sign switch
+            {
+                '+' => false,
+                '-' => true,
+                _   => throw new FormatException($"Invalid overpunch sign: '{sign}'"),
+            };
 
-            bytes[index] = opVal.Digit;
-            sign = opVal.Sign;
+            bytes[index] = (byte)opVal[1];
         }
 
         EnsureAllAsciiDigits(bytes);
@@ -45,12 +52,12 @@ public static class OpCodec
     /// <summary>
     /// 符號(sign)與數字文(numeric) → PIC 9/S9
     /// </summary>
-    /// <param name="sign">符號</param>
+    /// <param name="isNegative"></param>
     /// <param name="numeric">數字文 (char[])</param>
     /// <param name="pic"></param>
     /// <param name="options"></param>
     /// <returns></returns>
-    public static byte[] Encode(decimal sign, byte[] numeric, PicMeta pic, CodecOptions options)
+    public static byte[] Encode(bool isNegative, byte[] numeric, PicMeta pic, CodecOptions options)
     {
         EnsureAllAsciiDigits(numeric);
 
@@ -63,50 +70,15 @@ public static class OpCodec
                 _ => throw new FormatException($"Unsupported Sign option: {options.Sign}")
             };
 
-            byte digit = (byte)(numeric[index] & 0x7F); // ASCII overpunch
+            char digit = (char)(numeric[index] & 0x7F); // ASCII overpunch
+            string opValue = OpCodexBase.OpVal(isNegative, digit);
 
-            byte value = GetOpKey(new OpVal(sign, digit), options.DataStorage);
+            byte value = OpCodex.TryGetKey(opValue, options.DataStorage);
 
             numeric[index] = value;
         }
 
         return numeric;
-    }
-
-    /// <summary>
-    /// Get Overpunch Value
-    /// </summary>
-    /// <param name="key"></param>
-    /// <param name="ds">DataStorage Options</param>
-    /// <returns></returns>
-    /// <exception cref="FormatException"></exception>
-    private static OpVal GetOpValue(byte key, DataStorageOptions ds)
-    {
-        if (!OpCodex.Map.TryGetValue(ds, out Dictionary<byte, OpVal>? codex))
-            throw new FormatException($"Unsupported DataStorage: {ds}");
-
-        if (!codex.TryGetValue(key, out OpVal value))
-            throw new FormatException($"Invalid overpunch search key: '{key}'");
-
-        return value;
-    }
-
-    /// <summary>
-    /// Get Overpunch Key
-    /// </summary>
-    /// <param name="value"></param>
-    /// <param name="ds">DataStorage Options</param>
-    /// <returns></returns>
-    /// <exception cref="FormatException"></exception>
-    private static byte GetOpKey(OpVal value, DataStorageOptions ds)
-    {
-        if (!OpCodex.ReversedMap.TryGetValue(ds, out Dictionary<OpVal, byte>? codex))
-            throw new FormatException($"Unsupported DataStorage: {ds}");
-
-        if (!codex.TryGetValue(value, out byte key))
-            throw new FormatException($"Invalid overpunch search value: '{value}'");
-
-        return key;
     }
 
     private static void EnsureAllAsciiDigits(ReadOnlySpan<byte> span)